Some bugs were fixed for individual torrent settings

+++ TransmissionRPCClient/TRInfo.h
@@ -85,7 +85,7 @@
 @property(nonatomic) int           seedIdleMode;
 @property(nonatomic) int           seedIdleLimit;
 @property(nonatomic) int           seedRatioMode;
-@property(nonatomic) int           seedRatioLimit;
+@property(nonatomic) float         seedRatioLimit;
 @property(nonatomic) int           queuePosition;
 
 @end
